+++ ckanext/harvest/commands/harvester.py
@@ -1,8 +1,5 @@
 from __future__ import print_function
-
 import sys
-
-import six
 
 from ckan import model
 from ckan.logic import get_action, ValidationError
@@ -10,6 +7,7 @@ from ckan.logic import get_action, ValidationError
 from ckantoolkit import CkanCommand
 
 import ckanext.harvest.utils as utils
+from ckanext.harvest.logic.schema import unicode_safe
 
 
 class Harvester(CkanCommand):
@@ -34,11 +32,13 @@ class Harvester(CkanCommand):
         - clears all datasets, jobs and objects related to a harvest source,
           but keeps the source itself
 
-      harvester clearsource_history [{source-id}]
+      harvester clearsource_history [{source-id}] [-k]
         - If no source id is given the history for all harvest sources (maximum is 1000) will be cleared.
           Clears all jobs and objects related to a harvest source, but keeps the source itself.
           The datasets imported from the harvest source will NOT be deleted!!!
           If a source id is given, it only clears the history of the harvest source with the given source id.
+
+          To keep the currently active jobs use the -k option.
 
       harvester sources [all]
         - lists harvest sources
@@ -190,6 +190,14 @@ class Harvester(CkanCommand):
             will be aborted. You can use comma as a separator to provide multiple source_id's""",
         )
 
+        self.parser.add_option(
+            "-k",
+            "--keep-current",
+            dest="keep_current",
+            default=False,
+            help="Do not delete relevant harvest objects",
+        )
+
     def command(self):
         self._load_config()
 
@@ -265,23 +273,23 @@ class Harvester(CkanCommand):
     def create_harvest_source(self):
 
         if len(self.args) >= 2:
-            name = six.text_type(self.args[1])
+            name = unicode_safe(self.args[1])
         else:
             print("Please provide a source name")
             sys.exit(1)
         if len(self.args) >= 3:
-            url = six.text_type(self.args[2])
+            url = unicode_safe(self.args[2])
         else:
             print("Please provide a source URL")
             sys.exit(1)
         if len(self.args) >= 4:
-            type = six.text_type(self.args[3])
+            type = unicode_safe(self.args[3])
         else:
             print("Please provide a source type")
             sys.exit(1)
 
         if len(self.args) >= 5:
-            title = six.text_type(self.args[4])
+            title = unicode_safe(self.args[4])
         else:
             title = None
         if len(self.args) >= 6:
@@ -291,17 +299,17 @@ class Harvester(CkanCommand):
         else:
             active = True
         if len(self.args) >= 7:
-            owner_org = six.text_type(self.args[6])
+            owner_org = unicode_safe(self.args[6])
         else:
             owner_org = None
         if len(self.args) >= 8:
-            frequency = six.text_type(self.args[7])
+            frequency = unicode_safe(self.args[7])
             if not frequency:
                 frequency = "MANUAL"
         else:
             frequency = "MANUAL"
         if len(self.args) >= 9:
-            source_config = six.text_type(self.args[8])
+            source_config = unicode_safe(self.args[8])
         else:
             source_config = None
         try:
@@ -316,16 +324,17 @@ class Harvester(CkanCommand):
         print(result)
 
     def clear_harvest_source_history(self):
+        keep_current = bool(self.options.keep_current)
         source_id = None
         if len(self.args) >= 2:
-            source_id = six.text_type(self.args[1])
+            source_id = unicode_safe(self.args[1])
 
-        print(utils.clear_harvest_source_history(source_id))
+        print(utils.clear_harvest_source_history(source_id, keep_current))
 
     def show_harvest_source(self):
 
         if len(self.args) >= 2:
-            source_id_or_name = six.text_type(self.args[1])
+            source_id_or_name = unicode_safe(self.args[1])
         else:
             print("Please provide a source name")
             sys.exit(1)
@@ -333,7 +342,7 @@ class Harvester(CkanCommand):
 
     def remove_harvest_source(self):
         if len(self.args) >= 2:
-            source_id_or_name = six.text_type(self.args[1])
+            source_id_or_name = unicode_safe(self.args[1])
         else:
             print("Please provide a source id")
             sys.exit(1)
@@ -341,7 +350,7 @@ class Harvester(CkanCommand):
 
     def clear_harvest_source(self):
         if len(self.args) >= 2:
-            source_id_or_name = six.text_type(self.args[1])
+            source_id_or_name = unicode_safe(self.args[1])
         else:
             print("Please provide a source id")
             sys.exit(1)
@@ -357,7 +366,7 @@ class Harvester(CkanCommand):
 
     def create_harvest_job(self):
         if len(self.args) >= 2:
-            source_id_or_name = six.text_type(self.args[1])
+            source_id_or_name = unicode_safe(self.args[1])
         else:
             print("Please provide a source id")
             sys.exit(1)
@@ -368,7 +377,7 @@ class Harvester(CkanCommand):
 
     def job_abort(self):
         if len(self.args) >= 2:
-            job_or_source_id_or_name = six.text_type(self.args[1])
+            job_or_source_id_or_name = unicode_safe(self.args[1])
         else:
             print("Please provide a job id or source name/id")
             sys.exit(1)
@@ -383,7 +392,7 @@ class Harvester(CkanCommand):
         if len(self.args) >= 2:
             if len(self.args) >= 3 and self.args[2].startswith('force-import='):
                 force_import = self.args[2].split('=')[-1]
-            source_id_or_name = six.text_type(self.args[1])
+            source_id_or_name = unicode_safe(self.args[1])
         else:
             print("Please provide a source id")
             sys.exit(1)
@@ -393,7 +402,7 @@ class Harvester(CkanCommand):
     def import_stage(self):
 
         if len(self.args) >= 2:
-            source_id_or_name = six.text_type(self.args[1])
+            source_id_or_name = unicode_safe(self.args[1])
             context = {
                 "model": model,
                 "session": model.Session,
@@ -429,7 +438,7 @@ class Harvester(CkanCommand):
     def abort_failed_jobs(self):
         job_life_span = False
         if len(self.args) >= 2:
-            job_life_span = six.text_type(self.args[1])
+            job_life_span = unicode_safe(self.args[1])
 
         utils.abort_failed_jobs(
             job_life_span,