+++ ckanext/harvest/harvesters/base.py
@@ -3,7 +3,6 @@
 import logging
 import re
 import uuid
-import six
 
 from sqlalchemy import exists, and_
 from sqlalchemy.orm import contains_eager
@@ -24,6 +23,7 @@ from ckanext.harvest.model import (HarvestObject, HarvestGatherError,
 
 from ckan.plugins.core import SingletonPlugin, implements
 from ckanext.harvest.interfaces import IHarvester
+from ckanext.harvest.logic.schema import unicode_safe
 
 if p.toolkit.check_ckan_version(min_version='2.3'):
     from ckan.lib.munge import munge_tag
@@ -278,7 +278,7 @@ class HarvesterBase(SingletonPlugin):
         try:
             # Change default schema
             schema = default_create_package_schema()
-            schema['id'] = [ignore_missing, six.text_type]
+            schema['id'] = [ignore_missing, unicode_safe]
             schema['__junk'] = [ignore]
 
             # Check API version
@@ -321,6 +321,9 @@ class HarvesterBase(SingletonPlugin):
                     package_dict.setdefault('name',
                                             existing_package_dict['name'])
 
+                    for field in p.toolkit.aslist(config.get('ckan.harvest.not_overwrite_fields')):
+                        if field in existing_package_dict:
+                            package_dict[field] = existing_package_dict[field]
                     new_package = p.toolkit.get_action(
                         'package_update' if package_dict_form == 'package_show'
                         else 'package_update_rest')(context, package_dict)