+++ ckanext/harvest/helpers.py
@@ -56,7 +56,10 @@ def package_list_for_source(source_id):
 
     query = logic.get_action('package_search')(context, search_dict)
 
-    base_url = h.url_for('{0}_read'.format(DATASET_TYPE_NAME), id=source_id)
+    base_url = h.url_for(
+        '{0}_read'.format(DATASET_TYPE_NAME),
+        id=harvest_source['name']
+    )
 
     def pager_url(q=None, page=None):
         url = base_url
@@ -133,7 +136,7 @@ def harvest_source_extra_fields():
     for harvester in p.PluginImplementations(IHarvester):
         if not hasattr(harvester, 'extra_schema'):
             continue
-        fields[harvester.info()['name']] = harvester.extra_schema().keys()
+        fields[harvester.info()['name']] = list(harvester.extra_schema().keys())
     return fields