+++ ckanext/harvest/plugin/__init__.py
@@ -2,9 +2,10 @@
 
 import os
 import json
+import six
 from logging import getLogger
 
-from six import string_types, text_type
+from six import string_types
 from collections import OrderedDict
 
 from ckan import logic
@@ -61,102 +62,142 @@ class Harvest(MixinPlugin, p.SingletonPlugin, DefaultDatasetForm, DefaultTransla
 
     # IPackageController
 
+    # CKAN < 2.10 hooks
     def after_create(self, context, data_dict):
-        if 'type' in data_dict and data_dict['type'] == DATASET_TYPE_NAME and not self.startup:
+        return self.after_dataset_create(context, data_dict)
+
+    def after_update(self, context, data_dict):
+        return self.after_dataset_update(context, data_dict)
+
+    def after_delete(self, context, data_dict):
+        return self.after_dataset_delete(context, data_dict)
+
+    def before_search(self, search_params):
+        return self.before_dataset_search(search_params)
+
+    def before_index(self, pkg_dict):
+        return self.before_dataset_index(pkg_dict)
+
+    def after_show(self, context, data_dict):
+        return self.after_dataset_show(context, data_dict)
+
+    # CKAN >= 2.10 hooks
+    def after_dataset_create(self, context, data_dict):
+        if (
+            "type" in data_dict
+            and data_dict["type"] == DATASET_TYPE_NAME
+            and not self.startup
+        ):
             # Create an actual HarvestSource object
             _create_harvest_source_object(context, data_dict)
 
-    def after_update(self, context, data_dict):
-        if 'type' in data_dict and data_dict['type'] == DATASET_TYPE_NAME:
+    def after_dataset_update(self, context, data_dict):
+        if "type" in data_dict and data_dict["type"] == DATASET_TYPE_NAME:
             # Edit the actual HarvestSource object
             _update_harvest_source_object(context, data_dict)
 
-    def after_delete(self, context, data_dict):
+    def after_dataset_delete(self, context, data_dict):
 
-        package_dict = p.toolkit.get_action('package_show')(context, {'id': data_dict['id']})
+        package_dict = p.toolkit.get_action("package_show")(
+            context, {"id": data_dict["id"]}
+        )
 
-        if 'type' in package_dict and package_dict['type'] == DATASET_TYPE_NAME:
+        if "type" in package_dict and package_dict["type"] == DATASET_TYPE_NAME:
             # Delete the actual HarvestSource object
             _delete_harvest_source_object(context, package_dict)
 
-    def before_view(self, data_dict):
+    def before_dataset_search(self, search_params):
+        """Prevents the harvesters being shown in dataset search results."""
 
-        # check_ckan_version should be more clever than this
-        if p.toolkit.check_ckan_version(max_version='2.1.99') and (
-           'type' not in data_dict or data_dict['type'] != DATASET_TYPE_NAME):
-            # This is a normal dataset, check if it was harvested and if so, add
-            # info about the HarvestObject and HarvestSource
-            harvest_object = model.Session.query(HarvestObject) \
-                    .filter(HarvestObject.package_id == data_dict['id']) \
-                    .filter(HarvestObject.current==True).first() # noqa
-
-            if harvest_object:
-                for key, value in [
-                    ('harvest_object_id', harvest_object.id),
-                    ('harvest_source_id', harvest_object.source.id),
-                    ('harvest_source_title', harvest_object.source.title),
-                        ]:
-                    _add_extra(data_dict, key, value)
-        return data_dict
-
-    def before_search(self, search_params):
-        '''Prevents the harvesters being shown in dataset search results.'''
-
-        fq = search_params.get('fq', '')
-        if 'dataset_type:harvest' not in fq:
-            fq = u"{0} -dataset_type:harvest".format(search_params.get('fq', ''))
-            search_params.update({'fq': fq})
+        fq = search_params.get("fq", "")
+        if "dataset_type:harvest" not in fq:
+            fq = "{0} -dataset_type:harvest".format(fq.encode('utf8') if six.PY2 else fq)
+            search_params.update({"fq": fq})
 
         return search_params
 
-    def after_show(self, context, data_dict):
+    def before_dataset_index(self, pkg_dict):
 
-        if 'type' in data_dict and data_dict['type'] == DATASET_TYPE_NAME:
+        harvest_object = model.Session.query(HarvestObject) \
+            .filter(HarvestObject.package_id == pkg_dict["id"]) \
+            .filter(HarvestObject.current == True).first() # noqa
+
+        if harvest_object:
+
+            data_dict = json.loads(pkg_dict["data_dict"])
+
+            validated_data_dict = json.loads(pkg_dict["validated_data_dict"])
+
+            harvest_extras = [
+                ("harvest_object_id", harvest_object.id),
+                ("harvest_source_id", harvest_object.source.id),
+                ("harvest_source_title", harvest_object.source.title),
+            ]
+
+            for key, value in harvest_extras:
+
+                # If the harvest extras are there, update them. This can
+                # happen eg when calling package_update or resource_update,
+                # which call package_show
+                harvest_not_found = True
+                harvest_not_found_validated = True
+                if not data_dict.get("extras"):
+                    data_dict["extras"] = []
+
+                for e in data_dict.get("extras"):
+                    if e.get("key") == key:
+                        e.update({"value": value})
+                        harvest_not_found = False
+                if harvest_not_found:
+                    data_dict["extras"].append({"key": key, "value": value})
+
+                if not validated_data_dict.get("extras"):
+                    validated_data_dict["extras"] = []
+
+                for e in validated_data_dict.get("extras"):
+                    if e.get("key") == key:
+                        e.update({"value": value})
+                        harvest_not_found_validated = False
+                if harvest_not_found_validated:
+                    validated_data_dict["extras"].append({"key": key, "value": value})
+
+                # The commented line isn't cataloged correctly, if we pass the
+                # basic key the extras are prepended and the system works as
+                # expected.
+                # pkg_dict['extras_{0}'.format(key)] = value
+                pkg_dict[key] = value
+
+            pkg_dict["data_dict"] = json.dumps(data_dict)
+            pkg_dict["validated_data_dict"] = json.dumps(validated_data_dict)
+
+        if isinstance(pkg_dict.get('status'), dict):
+            try:
+                pkg_dict['status'] = json.dumps(pkg_dict['status'])
+            except ValueError:
+                pkg_dict.pop('status', None)
+
+        return pkg_dict
+
+    def after_dataset_show(self, context, data_dict):
+
+        if "type" in data_dict and data_dict["type"] == DATASET_TYPE_NAME:
             # This is a harvest source dataset, add extra info from the
             # HarvestSource object
-            source = HarvestSource.get(data_dict['id'])
+            source = HarvestSource.get(data_dict["id"])
             if not source:
-                log.error('Harvest source not found for dataset {0}'.format(data_dict['id']))
+                log.error(
+                    "Harvest source not found for dataset {0}".format(data_dict["id"])
+                )
                 return data_dict
 
-            st_action_name = 'harvest_source_show_status'
+            st_action_name = "harvest_source_show_status"
             try:
                 status_action = p.toolkit.get_action(st_action_name)
             except KeyError:
                 logic.clear_actions_cache()
                 status_action = p.toolkit.get_action(st_action_name)
 
-            data_dict['status'] = status_action(context, {'id': source.id})
-
-        elif 'type' not in data_dict or data_dict['type'] != DATASET_TYPE_NAME:
-            # This is a normal dataset, check if it was harvested and if so, add
-            # info about the HarvestObject and HarvestSource
-
-            harvest_object = model.Session.query(HarvestObject) \
-                    .filter(HarvestObject.package_id == data_dict['id']) \
-                    .filter(HarvestObject.current == True).first() # noqa
-
-            # If the harvest extras are there, remove them. This can happen eg
-            # when calling package_update or resource_update, which call
-            # package_show
-            if data_dict.get('extras'):
-                data_dict['extras'][:] = [e for e in data_dict.get('extras', [])
-                                          if not e['key']
-                                          in ('harvest_object_id', 'harvest_source_id', 'harvest_source_title',)]
-
-            # We only want to add these extras at index time so they are part
-            # of the cached data_dict used to display, search results etc. We
-            # don't want them added when editing the dataset, otherwise we get
-            # duplicated key errors.
-            # The only way to detect indexing right now is checking that
-            # validate is set to False.
-            if harvest_object and not context.get('validate', True):
-                for key, value in [
-                    ('harvest_object_id', harvest_object.id),
-                    ('harvest_source_id', harvest_object.source.id),
-                    ('harvest_source_title', harvest_object.source.title),
-                        ]:
-                    _add_extra(data_dict, key, value)
+            data_dict["status"] = status_action(context, {"id": source.id})
 
         return data_dict
 
@@ -192,11 +233,10 @@ class Harvest(MixinPlugin, p.SingletonPlugin, DefaultDatasetForm, DefaultTransla
         Returns the schema for mapping package data from a form to a format
         suitable for the database.
         '''
-        from ckanext.harvest.logic.schema import harvest_source_create_package_schema
+        from ckanext.harvest.logic.schema import harvest_source_create_package_schema, unicode_safe
         schema = harvest_source_create_package_schema()
         if self.startup:
-            schema['id'] = [text_type]
-
+            schema['id'] = [unicode_safe]
         return schema
 
     def update_package_schema(self):
@@ -244,8 +284,8 @@ class Harvest(MixinPlugin, p.SingletonPlugin, DefaultDatasetForm, DefaultTransla
         p.toolkit.add_resource('../public/ckanext/harvest/javascript', 'harvest-extra-field')
 
         if p.toolkit.check_ckan_version(min_version='2.9.0'):
-            mappings = config.get('ckan.legacy_route_mappings', {})
-            if isinstance(mappings, string_types):
+            mappings = config.get('ckan.legacy_route_mappings') or {}
+            if mappings and isinstance(mappings, string_types):
                 mappings = json.loads(mappings)
 
             mappings.update({
@@ -315,15 +355,6 @@ class Harvest(MixinPlugin, p.SingletonPlugin, DefaultDatasetForm, DefaultTransla
         return OrderedDict([('frequency', 'Frequency'),
                             ('source_type', 'Type'),
                             ])
-
-
-def _add_extra(data_dict, key, value):
-    if 'extras' not in data_dict:
-        data_dict['extras'] = []
-
-    data_dict['extras'].append({
-        'key': key, 'value': value, 'state': u'active'
-    })
 
 
 def _get_logic_functions(module_root, logic_functions={}):